Fold LoggerProviderBuilderBase into LoggerProviderBuilderSdk.
